Update search, delete, and UI functionality
Updated `SearchAuctionSessions` in `AuctionSessionController.cs` to filter by `searchStartTime` and `searchEndTime` instead of `searchPrice` and `sortOrder`. Modified `Delete` in `KoiFishController.cs` to redirect directly to `Index` with a message parameter. Changed title in `Index.cshtml` to "Auction Session List" and updated the search form to include `StartTime` and `EndTime` fields with AJAX functionality for dynamic updates. Cleaned up CSS in `Index.cshtml` and added a script for AJAX search. Added a check in `KoiFishService.cs` to prevent deletion of koi fish in an auction session and removed an unused `Azure.Core` import.

+++ KoiFishAuction.Service/Services/Implementation/KoiFishService.cs
@@ -1,4 +1,5 @@
-﻿using KoiFishAuction.Common.RequestModels.KoiFish;
+﻿using Azure.Core;
+using KoiFishAuction.Common.RequestModels.KoiFish;
 using KoiFishAuction.Common.ViewModels.KoiFish;
 using KoiFishAuction.Data;
 using KoiFishAuction.Data.Models;
@@ -83,6 +84,12 @@ namespace KoiFishAuction.Service.Services.Implementation
                 if (koiFish == null)
                 {
                     return new ServiceResult<bool>(Common.Constant.StatusCode.FailedStatusCode, "This koi fish does not exist.");
+                }
+
+                var result = await _unitOfWork.KoiFishRepository.IsKoiInAuction(id);
+                if (result)
+                {
+                    return new ServiceResult<bool>(Common.Constant.StatusCode.FailedStatusCode, "This koiFish is in an auction session and cannot delete.");
                 }
 
                 await _unitOfWork.KoiFishRepository.RemoveAsync(koiFish);
